langchain-python/agents/agent-coordinator/requirements.txt
chromadb==0.4.14
fastapi==0.105.0
fuzzywuzzy==0.18.0
google-api-python-client==2.114.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-aiplatform==1.39.0
google-cloud-datastore==2.8.0
itsdangerous==2.1.2
python-jose==3.3.0
langchain==0.1.3
langchain-community==0.0.15
langchain-core==0.1.15
langserve==0.0.41
langsmith==0.0.83
langchain-google-vertexai==0.0.2
pytz==2023.3.post1
sse-starlette==1.8.2
transformers==4.37.1
unidecode==1.3.8
uvicorn==0.23.2
